Gossip Partner/requirements.txt
SpeechRecognition
gtts
pygame
Pillow
pyaudio
